Refactor SensorListScreen, extract SensorListState
